feat(ci): opt-in doc-context injection for CI reviews
Extract doc-context scanning into a shared core module and wire it into the
CI runner behind an explicit, opt-in `doc-dirs` action input (env
PI_REVIEWER_DOC_DIRS). Empty by default — CI injects no docs unless the repo
maintainer configures dirs, avoiding surprise context/token cost. Interactive
extension keeps its opt-out defaults.

- src/core/doc-context.ts: framework-free loadDocContext() over FsOps
- extensions/pi-reviewer-doc-context: reuse shared core logic
- src/ci/review.ts: parseDocDirs + inject only when doc-dirs configured
- action.yml: doc-dirs input -> PI_REVIEWER_DOC_DIRS env
- src/ci/init.ts: surface doc-dirs as commented example in workflow
- tests: opt-in/env/parse coverage; updated init snapshot
