+++ CI.md
@@ -42,6 +42,9 @@ jobs:
           github-token: ${{ secrets.GITHUB_TOKEN }}
           pi-api-key: ${{ secrets.PI_API_KEY }}
           min-severity: ${{ inputs.min-severity || 'info' }}
+          # Opt in to injecting matching project docs into the review.
+          # Comma-separated dirs scanned for .md files with a 'description' frontmatter.
+          # doc-dirs: '.pi/notes,docs/review'
 ```
 
 Commit it to your default branch, then add your API key to your repo secrets:
@@ -60,6 +63,33 @@ Every pull request triggers an automatic review comment posted by `github-action
 | `model` | no | Model to use in `provider/modelId` format (e.g. `anthropic/claude-opus-4-6`) |
 | `post-comment` | no | Post review as a GitHub PR comment (default: `true`) |
 | `min-severity` | no | Minimum severity: `info`, `warn`, or `critical` (default: `info`) |
+| `doc-dirs` | no | Comma-separated dirs to scan for docs to inject into the review (default: empty — inject nothing) |
+
+## Doc context
+
+The reviewer can pull relevant project documentation into the review prompt based on which files changed in the diff. It is **opt-in** in CI: nothing is injected unless you set `doc-dirs`.
+
+```yaml
+      - uses: zeflq/pi-reviewer@main
+        with:
+          github-token: ${{ secrets.GITHUB_TOKEN }}
+          pi-api-key: ${{ secrets.PI_API_KEY }}
+          doc-dirs: '.pi/notes,docs/review'
+```
+
+For each configured dir, any `.md` file with a `description` frontmatter field is a candidate:
+
+```markdown
+---
+description: Authentication flows, JWT tokens, session management
+---
+
+# Auth Guide
+
+...content injected into the review when auth-related files change...
+```
+
+At review time, the action extracts keywords from the changed file paths (e.g. `src/auth/login.ts` → `auth`, `login`), then injects any doc whose `description` or filename matches a keyword. Keep descriptions specific enough to avoid over-matching, but broad enough to cover the files they apply to.
 
 ## Bot identity
