+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pi-reviewer",
-  "version": "0.5.3",
+  "version": "0.5.4",
   "private": true,
   "type": "module",
   "bin": {
